requirements.txt
langchain-core
langchain-openai
pydantic
python-dotenv
pypdf
matplotlib
graphviz
networkx
langchain
pytest
reportlab
pytest-asyncio